Updated color, text, content, and drop shadow --- OH MY!
Swapped background and content background-colors. Dark colors in
background and lighter colors in the foreground to bring content forward
and make it pop more. Made the font size larger.
Fixed some of the content to sound slightly better.
Added ordered lists to resumeCards.
Added drop shadow to the cards to give the page more depth